requirements.txt
python-telegram-bot==13.15
praw==7.7.1
APScheduler==3.6.3
python-dotenv==1.0.1
openai==1.14.3